Add Reception.java

+++ src/main/java/com/example/pokersc/entity/User.java
@@ -65,7 +65,7 @@ public class User extends Thread {
         this.total_round = total_round;
     }
 
-    public double getBuyin() {
+    public int getBuyin() {
         return buyin;
     }
